Вторая часть рефактора.
