+++ commands/adt_team/team_help_command.py
@@ -1,4 +1,5 @@
 import discord
+
 from bot_init import bot
 from config import ADMIN_TEAM
