+++ commands/github/achang_command.py
@@ -2,6 +2,7 @@ import discord
 from discord.ext import commands
 
 from bot_init import bot
+
 from .github_processor import (create_embed_list, fetch_github_data,
                                send_embeds, validate_repository, validate_user)
