+++ commands/github/check_workflows.py
@@ -1,4 +1,5 @@
 import sys
+
 import aiohttp
 
 from bot_init import bot
