+++ commands/restart_command.py
@@ -17,17 +17,12 @@ async def restart(ctx):
         # Уведомляем пользователя в текущем канале
         await ctx.send("🔄 Запущен протокол перезапуска. Пожалуйста, подождите.")
 
-        # Логируем событие перезапуска в канал логов
-        channel = bot.get_channel(LOG_CHANNEL_ID)
-        if channel:
-            await channel.send(
-                f"⚠️ {bot.user} завершает свою работу! Перезапуск начнётся в течение 10 минут."
-            )
-        else:
-            print(f"❌ Лог-канал с ID {LOG_CHANNEL_ID} не найден.")
-
         # Выполняем действия перед завершением работы
         await shutdown_def()
+
+        await ctx.send(
+            f"⚠️ {bot.user} завершает свою работу! Перезапуск начнётся в течение 10 минут."
+        )
 
         # Закрываем соединение бота
         await bot.close()
