+++ config.py
@@ -48,6 +48,12 @@ MESSAGE_ID_TIME_SHUTDOWS = 1320771150938243195
 TIME_SHUTDOWSE = 5 * 3600 + 57 * 60
 SS14_ADDRESS = "ss14://193.164.18.155"
 
+SS14_RUN_LEVELS = {
+    0: "Лобби",
+    1: "Раунд идёт",
+    2: "Окончание раунда..."
+}
+
 WHITELIST_ROLE_ID = [
     1060191651538145420,  # ID роли "Разработка"
     1116612861993689251,  # ID роли "Создатель проекта🔑"
