+++ events/on_ready.py
@@ -1,15 +1,17 @@
 import time
+import logging
 
 from bot_init import bot
 from commands.github import check_workflows
-from commands.github.git_fetch_pull import fetch_merged_pull_requests
-from commands.list_team_command import list_team_task
 from config import LOG_CHANNEL_ID
-from events.check_new_commit import monitor_commits
 from events.shutdows_after_time import shutdown_after_time
-from events.update_status import (update_status,
-                                  update_status_server_message_eddit)
-from events.update_time_shutdows import update_time_shutdows
+from tasks.check_new_commit_task import monitor_commits
+from tasks.git_fetch_pull_task import fetch_merged_pull_requests
+from tasks.list_team_task import list_team_task
+from tasks.update_status_presence_task import update_status_presence
+from tasks.update_status_server_message_eddit_task import \
+    update_status_server_message_eddit
+from tasks.update_time_shutdows_task import update_time_shutdows
 
 
 async def start_task_if_not_running(task, task_name: str):
@@ -28,6 +30,10 @@ async def on_ready():
     """
     Событие, которое выполняется при запуске бота.
     """
+    logging.info(f"Bot {bot.user.name} (ID: {bot.user.id}) is ready to work!")
+    logging.info("Connected to Discord successfully.")
+    logging.info(f"Guilds: {[guild.name for guild in bot.guilds]}")  # Выводит список серверов, к которым подключен бот.
+    
     bot.start_time = time.time()  # Сохраняем время старта бота
 
     # Проверка workflows на случай повторного запуска на GitHub Actions
@@ -37,7 +43,7 @@ async def on_ready():
     await start_task_if_not_running(fetch_merged_pull_requests, "fetch_merged_pull_requests")
     await start_task_if_not_running(list_team_task, "list_team_task")
     await start_task_if_not_running(monitor_commits, "monitor_commits")
-    await start_task_if_not_running(update_status, "update_status")
+    await start_task_if_not_running(update_status_presence, "update_status_presence")
     await start_task_if_not_running(update_status_server_message_eddit, "update_status_server_message_eddit")
     await start_task_if_not_running(update_time_shutdows, "update_time_shutdows")
 
@@ -46,7 +52,12 @@ async def on_ready():
     # Уведомляем в лог-канале, что бот активен
     channel = bot.get_channel(LOG_CHANNEL_ID)
     if channel:
-        await channel.send(f"{bot.user} активен!")
+        await channel.send(
+            content=(
+                f"✅ **{bot.user.name} успешно активирована!**\n"
+                f"🔹 **Статус:** Бот запущен и готов к работе.\n"
+            )
+        )
     else:
         print(f"❌ Не удалось найти канал с ID {LOG_CHANNEL_ID} для логов.")
