+++ events/update_status.py
@@ -16,56 +16,6 @@ SS14_RUN_LEVELS = {
     2: "Окончание раунда..."
 }
 
-@tasks.loop(seconds=15)
-async def update_status():
-    """
-    Обновление Rich Presence бота.
-    """
-    status_data = await get_ss14_server_status_second(SS14_ADDRESS)
-    if not status_data:
-        await bot.change_presence(activity=discord.Game(name="Ошибка при получении статуса"))
-        return
-
-    # Формируем строку для статуса
-    count = status_data.get("players", "?")
-    countmax = status_data.get("soft_max_players", "?")
-    name = status_data.get("name", "Неизвестно")
-    preset = status_data.get("preset", "?")
-    round_id = status_data.get("round_id", "?")
-    run_level = SS14_RUN_LEVELS.get(status_data.get("run_level"), "Неизвестно")
-
-    status_state = f"Игроков: {count}/{countmax} | Режим: {preset} | Раунд: {round_id} | Статус: {run_level}"
-    activity = discord.Activity(type=discord.ActivityType.playing, name=name, state=status_state)
-    await bot.change_presence(activity=activity)
-
-@tasks.loop(seconds=50)
-async def update_status_server_message_edit():
-    """
-    Фоновая задача для обновления статуса сервера в сообщении.
-    """
-    channel = bot.get_channel(CHANNEL_ID_UPDATE_STATUS)
-    if channel is None:
-        print("Не удалось найти канал!")
-        return
-
-    try:
-        message = await channel.fetch_message(1320771122433622084)
-        status_data = await get_ss14_server_status_second(SS14_ADDRESS)
-
-        if not status_data:
-            embed = create_error_embed(SS14_ADDRESS)
-        else:
-            embed = create_status_embed(SS14_ADDRESS, status_data)
-
-        await message.edit(embed=embed)
-
-    except discord.NotFound:
-        print("Сообщение не найдено!")
-    except discord.Forbidden:
-        print("У бота нет прав для редактирования сообщения!")
-    except Exception as e:
-        print(f"Ошибка при обновлении сообщения: {e}")
-
 async def get_ss14_server_status_second(address: str) -> dict:
     """
     Получает статус игры с сервера SS14.
